app/requirements.txt
Flask
psycopg2-binary
prometheus_client
